Rename vMerger formatters.cpp
