+++ source/CMakeLists.txt
@@ -1,11 +1,10 @@
 include(fetch_dragonbox)
 
 target_sources(nanofmt PRIVATE
-    "formatters.cpp"
+    "format.cpp"
     "numeric_utils.h"
     "parse_utils.h"
     "to_chars.cpp"
-    "vformat.cpp"
 )
 
 target_link_libraries(nanofmt PRIVATE dragonbox::dragonbox_to_chars)
